LayoutTests/fast/dom/timer-in-microtask-does-not-inherit-nesting-level.html

<!DOCTYPE html>
<html>
<body>
<script src="../../resources/js-test.js"></script>
<script>
jsTestIsAsync = true;
description("Tests that a timer that is scheduled during a microtask checkpoint does not inherit the nesting level of the timer task that just finished running.");

if (!window.testRunner || !window.internals) {
    testFailed("Test requires internals.");
    finishJSTest();
}

setTimeout(() => {
    setTimeout(() => {
        shouldBe("internals.timerNestingLevel()", "2");
        queueMicrotask(() => {
            shouldBe("internals.timerNestingLevel()", "0");     // not 2
            setTimeout(() => {
                shouldBe("internals.timerNestingLevel()", "1"); // not 3
                setTimeout(() => {
                    shouldBe("internals.timerNestingLevel()", "2"); // timers should still nest in a microtask context.
                    finishJSTest();
                }, 10);
            }, 10);
        });
    }, 10);
}, 10);
</script>
</body>
</html>
