LayoutTests/fast/dom/timer-in-microtask-does-not-inherit-nesting-level-expected.txt

Tests that a timer that is scheduled during a microtask checkpoint does not inherit the nesting level of the timer task that just finished running.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS internals.timerNestingLevel() is 2
PASS internals.timerNestingLevel() is 0
PASS internals.timerNestingLevel() is 1
PASS internals.timerNestingLevel() is 2
PASS successfullyParsed is true

TEST COMPLETE
